SAAS-369 remove a duplicate test file
